refactor: move streaming route to chat route
