+++ frontend/src/routes/chat/[roomID]/+page.svelte
@@ -1,207 +1,316 @@
 <script lang="ts">
   import { onMount, onDestroy } from 'svelte';
   import { PUBLIC_SERVER_WS_URL } from '$env/static/public';
-  import { ANSWER, ICE_CANDIDATE, STREAMING, TRANSCRIPTION } from '@/lib/constants/constants';
-  import type { AnalyzedMessage } from '@/lib/constants/types';
+  import { v4 } from 'uuid';
+  import { OFFER, ANSWER, ICE_CANDIDATE, HANG_UP } from '@/lib/constants/constants';
+  import type {
+    StreamingOfferMessage,
+    StreamingAnswerMessage,
+    StreamingIceCandidateMessage,
+    AnalyzedMessage
+  } from '@/lib/constants/types';
 
+  import HangUp from '@/lib/components/HangUp.svelte';
+  import Emojis from '@/lib/components/Emojis.svelte';
+  import Transcription from '@/lib/components/Transcription.svelte';
+  import MicrophoneSelector from '@/lib/components/MicrophoneSelector.svelte';
   import { page } from '$app/state';
+  import { goto } from '$app/navigation';
+
   let roomID = page.params.roomID;
+  let userVideo: HTMLVideoElement;
+  let otherVideo: HTMLVideoElement;
+
+  let ws: WebSocket;
+  let peerConnection: RTCPeerConnection;
+  let localStream: MediaStream;
+
+  let connectedUsers = $state(1);
+
+  let selectedMicrophone = $state('default');
+  let isMicOn = $state(true);
+  let isVideoOn = $state(true);
+  let isClosedCaptionOn = $state(false);
+  let isBackHandOn = $state(true);
+
+  let showHangUpModal = $state(false);
+  let showMicrophoneModal = $state(false);
+  let showEmojiModal = $state(false);
+  let receivedEmoji = $state('');
 
   // Transcription
-  let wsTranscription: WebSocket | null = null;
-  let pcTranscription: RTCPeerConnection | null = null;
-  let streamTranscription: MediaStream | null = null;
-
   let messages: AnalyzedMessage[] = $state([]);
-  let isStreaming = false;
 
-  let microphoneList: MediaDeviceInfo[] = $state([]);
-  let selectedMicrophone: string = $state('default');
+  onMount(() => {
+    console.log(otherVideo?.srcObject);
+    // Connect to the signaling server
+    ws = new WebSocket(`${PUBLIC_SERVER_WS_URL}/join?roomID=${roomID}&userID=${v4()}`);
 
-  onMount(async () => {
-    navigator.mediaDevices.enumerateDevices().then((devices) => {
-      microphoneList = devices.filter((device) => device.kind === 'audioinput');
-    });
+    ws.onopen = () => {
+      console.log('WebSocket connection established');
+    };
 
-    startTranscriptionConnection();
+    ws.onmessage = async (event) => {
+      const message = JSON.parse(event.data);
+      console.log('Received signaling message:', message);
+
+      // Handle different message types
+      if (message.type === OFFER) {
+        let data: StreamingOfferMessage = message;
+        await handleOffer(data.payload);
+      } else if (message.type == ANSWER) {
+        let data: StreamingAnswerMessage = message;
+        await handleAnswer(data.payload);
+      } else if (message.type == ICE_CANDIDATE) {
+        let data: StreamingIceCandidateMessage = message;
+        await handleIceCandidate(data.payload);
+      } else if (message.type == HANG_UP) {
+        connectedUsers = 1;
+      } else if (message.type == 'emoji') {
+        receivedEmoji = message.payload;
+      }
+    };
+
+    ws.onclose = () => {
+      console.log('WebSocket connection closed');
+    };
+
+    // Open the camera and set up local video
+    openCamera();
   });
-
-  function setSelectedMicrophone() {
-    console.log(selectedMicrophone);
-    restartConnection();
-  }
 
   onDestroy(() => {
-    cleanup();
+    stopMediaTracks();
   });
 
-  async function startTranscriptionConnection() {
+  async function openCamera() {
     try {
-      wsTranscription = new WebSocket(`${PUBLIC_SERVER_WS_URL}/ws`);
+      localStream = await navigator.mediaDevices.getUserMedia({
+        video: true,
+        audio: true
+      });
+      userVideo.srcObject = localStream;
 
-      wsTranscription.onopen = async () => {
-        console.log('wsTranscription connected');
+      // Create the PeerConnection
+      createPeerConnection();
 
-        try {
-          pcTranscription = new RTCPeerConnection();
-
-          pcTranscription.onicecandidate = (event) => {
-            if (event.candidate) {
-              let answer: IceCandidateMessage = {
-                type: 'iceCandidate',
-                candidate: event.candidate
-              };
-              wsTranscription?.send(JSON.stringify(answer));
-            }
-          };
-
-          streamTranscription = await navigator.mediaDevices.getUserMedia({
-            audio: {
-              channelCount: 1,
-              sampleRate: 48000,
-              deviceId: selectedMicrophone
-            }
-          });
-          if (streamTranscription) {
-            streamTranscription.getTracks().forEach((track) => {
-              pcTranscription?.addTrack(track);
-            });
-          }
-
-          pcTranscription.onconnectionstatechange = async () => {};
-
-          pcTranscription.oniceconnectionstatechange = () => {
-            if (pcTranscription?.iceConnectionState === 'failed') {
-              console.log('ICE connection failed, restarting');
-              restartConnection();
-            }
-          };
-
-          const offerDescription = await pcTranscription.createOffer();
-          await pcTranscription.setLocalDescription(offerDescription);
-
-          if (offerDescription.sdp) {
-            let offer: OfferMessage = {
-              type: 'offer',
-              sdp: offerDescription.sdp
-            };
-            wsTranscription?.send(JSON.stringify(offer));
-          }
-        } catch (getUserMediaError) {
-          console.error('Error getting user media:', getUserMediaError);
-          cleanup();
-          return;
-        }
-
-        if (!wsTranscription) {
-          return;
-        }
-        wsTranscription.onmessage = async (event) => {
-          const message = JSON.parse(event.data);
-          if (message.type === ANSWER) {
-            let answer: AnswerMessage = {
-              type: ANSWER,
-              sdp: message.sdp
-            };
-            // @ts-ignore
-            const remoteDesc = new RTCSessionDescription(answer);
-            try {
-              await pcTranscription?.setRemoteDescription(remoteDesc);
-            } catch (setRemoteDescError) {
-              console.error('Error setting remote description:', setRemoteDescError);
-              cleanup();
-            }
-          } else if (message.type === ICE_CANDIDATE) {
-            try {
-              if (message.candidate) {
-                // Correctly parse the JSON string to an object
-                const candidate = JSON.parse(message.candidate);
-                await pcTranscription?.addIceCandidate(new RTCIceCandidate(candidate));
-              }
-            } catch (e) {
-              console.error('Error adding ICE candidate:', e);
-            }
-          } else if (message.type === STREAMING) {
-            if (message.isStreaming) {
-              console.log('Starting streaming');
-            } else {
-              console.log('Stopping streaming');
-            }
-          } else if (message.type === TRANSCRIPTION) {
-            var newMessage: AnalyzedMessage = {
-              uuid: message.uuid,
-              text: message.text,
-              profanityScore: message.profanity_score
-            };
-
-            const updatedMessages = [...messages, newMessage];
-            messages = updatedMessages.slice(-25);
-          }
-        };
-
-        wsTranscription.onclose = () => {
-          console.log('WebSocket closed');
-        };
-
-        wsTranscription.onerror = (error) => {
-          console.error('WebSocket error:', error);
-        };
-      };
-
-      wsTranscription.onerror = (error) => {
-        console.error('WebSocket connection error:', error);
-      };
+      // Add local stream tracks to the PeerConnection
+      localStream.getTracks().forEach((track) => {
+        peerConnection.addTrack(track, localStream);
+      });
     } catch (error) {
-      console.error('Error starting connection:', error);
-      cleanup();
+      console.error('Error accessing media devices:', error);
     }
   }
 
-  async function startStreaming() {
-    if (wsTranscription) {
-      isStreaming = !isStreaming;
-      let message: StreamingMessage = {
-        type: 'streaming',
-        isStreaming: isStreaming
+  function createPeerConnection() {
+    peerConnection = new RTCPeerConnection();
+
+    // Handle ICE candidates
+    peerConnection.onicecandidate = (event) => {
+      if (event.type === 'icecandidate' && event.candidate) {
+        const message: any = {
+          type: ICE_CANDIDATE,
+          payload: event.candidate
+        };
+        ws.send(JSON.stringify(message));
+      }
+    };
+
+    // Handle incoming tracks
+    peerConnection.ontrack = (event) => {
+      console.log('Received remote track:', event.streams);
+      if (event.streams.length > 0) {
+        otherVideo.srcObject = event.streams[0];
+        connectedUsers = 2;
+      }
+    };
+
+    // Handle negotiation
+    peerConnection.onnegotiationneeded = async () => {
+      const offer = await peerConnection.createOffer();
+      await peerConnection.setLocalDescription(offer);
+      if (offer.type === OFFER && offer) {
+        const offerMessage: any = {
+          type: OFFER,
+          payload: offer
+        };
+        ws.send(JSON.stringify(offerMessage));
+      }
+    };
+
+    // Handle connection state changes
+    peerConnection.onconnectionstatechange = () => {
+      console.log('Connection state:', peerConnection.connectionState);
+      if (peerConnection.connectionState === 'disconnected') {
+        connectedUsers = 1;
+      }
+    };
+  }
+
+  // handleOffer sets the remote description of the peer connection and creates an answer
+  async function handleOffer(offer: RTCSessionDescriptionInit) {
+    await peerConnection.setRemoteDescription(offer);
+    const answer = await peerConnection.createAnswer();
+    await peerConnection.setLocalDescription(answer);
+
+    if (answer.type === ANSWER && answer.sdp) {
+      const answerMessage: any = {
+        type: ANSWER,
+        payload: answer
       };
-      wsTranscription.send(JSON.stringify(message));
-    } else {
-      console.error('WebSocket not connected');
+      ws.send(JSON.stringify(answerMessage));
     }
   }
 
-  function restartConnection() {
-    cleanup();
-    startTranscriptionConnection();
+  // handleAnswer sets the remote description of the peer connection
+  async function handleAnswer(answer: RTCSessionDescriptionInit) {
+    try {
+      await peerConnection.setRemoteDescription(answer);
+    } catch (error) {
+      console.error('Error handling answer:', error);
+    }
   }
 
-  function cleanup() {
-    wsTranscription?.close();
-    pcTranscription?.close();
-    streamTranscription?.getTracks().forEach((track) => track.stop());
+  // handleIceCandidate adds the received ICE candidate to the peer connection
+  async function handleIceCandidate(candidate: RTCIceCandidateInit) {
+    try {
+      await peerConnection.addIceCandidate(candidate);
+    } catch (error) {
+      console.error('Error adding received ICE candidate:', error);
+    }
   }
+
+  const toggleCamera = () => {
+    const tracks = localStream.getVideoTracks();
+    tracks[0].enabled = !tracks[0].enabled;
+    isVideoOn = tracks[0].enabled;
+  };
+
+  $effect(() => {
+    console.log('Mic status:', isMicOn);
+    if (localStream) {
+      const tracks = localStream.getAudioTracks();
+      tracks[0].enabled = !tracks[0].enabled;
+      isMicOn = tracks[0].enabled;
+    }
+  });
+
+  $effect(() => {
+    console.log('Selected microphone:', selectedMicrophone);
+  });
+
+  const stopMediaTracks = () => {
+    ws?.close();
+    peerConnection?.close();
+    localStream?.getTracks().forEach((track) => track.stop());
+  };
+
+  const handleHangUp = () => {
+    ws.send(JSON.stringify({ type: HANG_UP }));
+    showHangUpModal = false;
+    stopMediaTracks();
+    goto('/');
+  };
+
+  const shareEmoji = (emoji: string) => {
+    ws.send(JSON.stringify({ type: 'emoji', payload: emoji }));
+  };
+
+  const handleClosedCaption = () => {
+    isClosedCaptionOn = !isClosedCaptionOn;
+  };
 </script>
 
-<div class="w-full justify-center text-center">
-  <div class="m-4 rounded bg-[darkgray] p-4">
-    <button onclick={startStreaming} aria-label="Start WebRTC">Start WebRTC</button>
-  </div>
-  <div class="m-4 text-left text-gray-600">
-    <span class="font-normal"> Transcription: </span>
-    {#each messages as message}
-      <span class={message.profanityScore > 0.9 ? 'text-red-500 line-through' : ''}>
-        {message.text}
-      </span>
-    {/each}
-  </div>
-  <div>
-    <select bind:value={selectedMicrophone} onchange={setSelectedMicrophone}>
-      {#each microphoneList as microphone}
-        <option value={microphone.deviceId}>{microphone.label}</option>
+<main class="flex min-h-screen flex-col items-center justify-center bg-[rgb(25,25,25)] p-24">
+  <div class="relative w-full max-w-5xl">
+    <div class="relative aspect-video w-full overflow-hidden rounded-lg bg-gray-900">
+      <div
+        class={`h-full w-full ${connectedUsers === 2 ? 'opacity-100' : 'opacity-0'} transition-opacity duration-500`}
+      >
+        <div class="h-full w-full overflow-hidden rounded-lg border-2 border-white">
+          <video
+            class="h-full w-full object-cover"
+            autoPlay
+            playsInline
+            muted
+            bind:this={otherVideo}
+          >
+            Your browser does not support the video tag.
+          </video>
+        </div>
+      </div>
+      <div
+        class={`absolute ${connectedUsers === 2 ? 'bottom-4 right-4 z-30 aspect-video w-1/4' : 'inset-0 z-10'} transition-all duration-500 ease-in-out`}
+      >
+        <div
+          class={`h-full w-full overflow-hidden rounded-lg ${
+            connectedUsers === 2 ? 'border-2 border-white' : 'border-2 border-white shadow-lg'
+          }`}
+        >
+          <video
+            class="h-full w-full object-cover"
+            autoPlay
+            muted
+            playsInline
+            bind:this={userVideo}
+          >
+            Your browser does not support the video tag.
+          </video>
+        </div>
+      </div>
+    </div>
+    <div class="m-4 text-left text-gray-600">
+      <span class="font-normal"> Transcription: </span>
+      {#each messages as message}
+        <span class={message.profanityScore > 0.9 ? 'text-red-500 line-through' : ''}>
+          {message.text}
+        </span>
       {/each}
-    </select>
+    </div>
+    <div class="relative mt-4 flex w-full justify-center space-x-4">
+      <MicrophoneSelector
+        bind:isMicOn
+        {showMicrophoneModal}
+        bind:selectedMicrophone
+        displayTop={true}
+      />
+
+      <button
+        title="Toggle Camera"
+        onclick={toggleCamera}
+        class={`my-auto flex select-none items-center justify-center rounded-full p-3 no-underline hover:brightness-75 ${isVideoOn ? 'bg-gray-200 dark:text-black' : 'bg-red-500'}`}
+      >
+        <span class="material-symbols-outlined"> videocam </span>
+      </button>
+      <button
+        onclick={handleClosedCaption}
+        class={`my-auto flex select-none items-center justify-center rounded-full p-3 no-underline hover:brightness-75 ${isClosedCaptionOn ? 'bg-green-500' : ' bg-gray-200 dark:text-black'}`}
+      >
+        <span class="material-symbols-outlined"> closed_caption </span>
+      </button>
+
+      <Emojis {showEmojiModal} {shareEmoji} {receivedEmoji} />
+
+      <!-- <button
+        onclick={() => (isBackHandOn = !isBackHandOn)}
+        class={`my-auto flex select-none items-center justify-center rounded-full p-3 no-underline hover:brightness-75 ${isBackHandOn ? 'bg-gray-200 dark:text-black' : ''}`}
+      >
+        <span class="material-symbols-outlined"> back_hand </span>
+      </button>
+      <button
+        onclick={() => console.log('more option')}
+        class={`my-auto flex select-none items-center justify-center rounded-full bg-gray-200 px-2 py-3 no-underline hover:brightness-75`}
+      >
+        <span class="material-symbols-outlined"> more_vert </span>
+      </button> -->
+
+      <!-- Close meeting -->
+      <HangUp {handleHangUp} {showHangUpModal} displayTop={true} />
+
+      {#if isClosedCaptionOn}
+        <Transcription {roomID} {selectedMicrophone} bind:messages></Transcription>
+      {/if}
+    </div>
   </div>
-  <div>
-    {selectedMicrophone}
-  </div>
-</div>
+</main>
